A function for estimating the degree of dosage compensation at X-linked loci using the delta, Fieller's or penalized Fieller's method

This function is used to calculate the upper and lower limits of the delta, Fieller's or penalized Fieller's confidence interval for the ratio estimate.

Frequen_DC(SNP,Y,trait_type='quantitative',Sex,Covariate=NULL,
           variance_heterogeneity=FALSE,MGC_Cutoff=20,method='Fieller',
           df_n=NULL,df_d=NULL,con_level=0.95,truncation_interval=FALSE,
           truncation_lower=NULL,truncation_upper=NULL,description=FALSE,
           H0_test=FALSE,DC_0)

Arguments

SNP

Each genotype is coded as 0, 1 or 2 for females and coded as 0 or 1 for males, indicating the number of reference allele. The length of SNP should be the same to those of Y, Sex and Covariate.

Y

A numeric vector of a quantitative trait, such as human height.

trait_type

A character string either being "quantitative" or "qualitative", the default value is "quantitative".

Sex

A vector of the genetic sex following PLINK default coding, where males are coded as 1 and females are coded as 2.

Covariate

Optional: a vector or a matrix of covariates, such as age and BMI.

variance_heterogeneity

For quantitative traits, we empolyed weighted least squares to fit the model when variance_heterogeneity is set to TRUE, and ordinary least squares when variance_heterogeneity is set to FASLE.

MGC_Cutoff

Cutoff for the minimum genotype count in either females or males (default=20). When the minimum genotype count for a SNP is below this specified cutoff, we will employ ordinary least squares instead of weighted least squares for genetic association analysis with quantitative traits. This is mainly to avoid inflated type I errors caused by the instability of the weights. (Yang et al., 2022; Deng et al., 2019; Soave et al., 2015).

method

a character string indicating which kind of estimation methods is to be used. One of "Fieller"(default), "PenFieller" , "delta" or "all": can be abbreviated.

df_n

The degree of freedom of the numerator. The default value is NULL.

df_d

The degree of freedom of the denominator. The default value is NULL.

con_level

The confidence level. Should be between 0 and 1. The default is 0.95.

truncation_interval

According to the literature or known knowledge, the ratio should be within a certain interval, for example in the interval from 1 to 2, so we need to truncate the point estimate and the Fieller's confidence interval into the certain interval to get the final results. The default value is FALSE, that is, no truncation.

truncation_lower

If truncation_interval=TRUE, you need to specify the truncation interval, and truncation_lower indicates the lower limit of the truncation interval. The default value is NULL.

truncation_upper

If truncation_interval=TRUE, you need to specify the truncation interval, and truncation_upper indicates the upper limit of the truncation interval. The default value is NULL.

description

Whether to describe the results of point_estimate and confidence interval in text. The default value is FASLE.

H0_test

Whether to test the null hypothesis of no DC, incomplete DC or complete DC. The default value is FASLE.

DC_0

If H0_test=TRUE, you need to specify the null hypothesis, where DC_0 takes the values of 1 and 2 for no DC or complete DC, respectively. The default value is NULL.

Value

a list:

point_estimate

The point estimate of the ratio.

Lower_CL

The lower bound of the estimated interval of the Fieller's method.

Upper_CL

The upper bound of the estimated interval of the Fieller's method.

CI

The estimated interval of the Fieller's method.

length_CI

The length of the estimated confidence interval.

CI_type

The CI of the Fieller's method can be divided into 'continuous interval','discontinuous interval', 'empty set', 'noninformative interval' or 'point'.

H0_test

If H0_test=TRUE, returns the Z-score test statistic and corresponding p-value under the null hypothesis.

References

Yang ZY, Liu W, Yuan YX, et al. Robust association tests for quantitative traits on the X chromosome. Heredity, 2022, 129: 244–256.

Deng WQ, Mao S, Kalnapenkis A, et al. Analytical strategies to include the X-chromosome in variance heterogeneity analyses: evidence for trait-specific polygenic variance structure. Genetic Epidemiology, 2019, 43: 815-830.

Soave D, Corvol H, Panjwani N, et al. A joint location-scale test improves power to detect associated SNPs, gene sets, and pathways. The American Journal of Human Genetics, 2015, 97: 125–138.

Examples

data("Graves_Meta_analysis")
Frequen_DC(Graves_Meta_analysis$rs3827440,Graves_Meta_analysis$Graves_disease,
           trait_type='qualitative',Graves_Meta_analysis$Sex,method='all',
           con_level=0.95,truncation_interval=TRUE,truncation_lower=1,truncation_upper=2,
           description=TRUE,H0_test=TRUE,DC_0=2)
#> $DC_Fieller
#> $DC_Fieller$Estimation_result
#>            point_estimate Lower_CL Upper_CL
#> origin           2.080558 1.359631 3.072538
#> truncation       2.000000 1.359631 2.000000
#>                                              CI length_CI             CI_type
#> origin     [1.35963051142375, 3.07253812022679] 1.7129076 continuous interval
#> truncation                [1.35963051142375, 2] 0.6403695 continuous interval
#> 
#> $DC_Fieller$H0_test
#>          Z_score   P_value
#> DC_0=2 -0.195643 0.8448896
#> 
#> 
#> $DC_PenFieller
#> $DC_PenFieller$Estimation_result
#>            point_estimate Lower_CL Upper_CL
#> origin             2.0801  1.35732  3.06474
#> truncation         2.0000  1.35732  2.00000
#>                                              CI length_CI             CI_type
#> origin     [1.35732045715608, 3.06473958535594] 1.7074191 continuous interval
#> truncation                [1.35732045715608, 2] 0.6426795 continuous interval
#> 
#> $DC_PenFieller$H0_test
#>           Z_score   P_value
#> DC_0=2 -0.1945366 0.8457557
#> 
#> 
#> $DC_delta
#> $DC_delta$Estimation_result
#>            point_estimate Lower_CL Upper_CL                                  CI
#> origin           2.080558 1.261162 2.899955 [1.2611616626819, 2.89995452508097]
#> truncation       2.000000 1.261162 2.000000                [1.2611616626819, 2]
#>            length_CI             CI_type
#> origin     1.6387929 continuous interval
#> truncation 0.7388383 continuous interval
#> 
#> $DC_delta$H0_test
#>          Z_score   P_value
#> DC_0=2 0.1926918 0.8472004
#> 
#>