Graves_Meta_analysis Dataset

A data frame containing X chromosome analysis for Graves’ disease

Graves_Meta_analysis

Format

a data frame with 11013 rows and 4 variables:

Graves_disease

Graves’ disease cases are coded as 1 and controls 0.

rs3827440

for females, each genotype is coded as 0, 1 and 2 for TT, TC and CC; for males, each genotype is coded as 0 and 1 for T and C.

Sex

the genetic sex of individuals in the sample population, where males are coded as 1 and females 2.

dataset

there were two stages of the association analysis in Chu et al. (J Med Genet 2013, 50: 479–85), i.e., the genome-wide association study (GWAS) stage and the replication stage

Source

According to Chu et al. (J Med Genet 2013, 50: 479–85), SNP rs3827440 within the GPR174 gene on X chromosome was detected to be associated with the Graves’ disease. There were two stages of the association analysis in Chu et al. (J Med Genet 2013, 50: 479–85), i.e., the genome-wide association study (GWAS) stage and the replication stage. The association between SNP rs3827440 and the Graves’ disease was identified in both of two stages and the pooled data of these two stages. There are 2941 subjects (699 males and 2242 females) in the GWAS stage and 8074 subjects (1814 males and 6260 females) in the replication stage. According to the summary statistics in Table 1 in Chu et al. (J Med Genet 2013, 50: 479–85), we can reconstruct this dataset.

References

Chu X, Shen M, Xie F, et al. An X chromosome-wide association analysis identifies variants in GPR174 as a risk factor for Graves' disease. Journal of medical genetics, 2013, 50: 479-485.